build_deploy_scripts/requirements.txt
watchdog
requests==2.31.0
pyyaml==6.0.1
fastapi==0.111.0